# backend/app/api/preferences.py
from fastapi import APIRouter, Depends
from pydantic import BaseModel
from sqlalchemy.orm import Session

from app.auth import require_permission
from app.database import get_db
from app.models.user import User
from app.schemas.common import MessageResponse
from app.serializers import serialize_message_response
from app.schemas.preferences import PreferencesResponse
from app.services.memory_service import MemoryService

router = APIRouter()


class SetPreferenceRequest(BaseModel):
    key: str
    value: str


class BatchPreferenceRequest(BaseModel):
    signature_org: str = ""
    default_tone: str = "formal"
    default_recipients: str = ""
    avoid_phrases: str = ""


@router.get("", response_model=PreferencesResponse)
def get_preferences(
    db: Session = Depends(get_db),
    current_user: User = Depends(require_permission("preferences:read")),
):
    svc = MemoryService(db, account_id=current_user.account_id)
    return svc.get_preferences(user_id=current_user.id)


@router.put("", response_model=MessageResponse)
def set_preference(
    req: SetPreferenceRequest,
    db: Session = Depends(get_db),
    current_user: User = Depends(require_permission("preferences:write")),
):
    svc = MemoryService(db, account_id=current_user.account_id)
    svc.set_preference(user_id=current_user.id, key=req.key, value=req.value, commit=False)
    db.commit()
    return serialize_message_response("偏好已保存")


@router.put("/batch", response_model=MessageResponse)
def set_preferences_batch(
    req: BatchPreferenceRequest,
    db: Session = Depends(get_db),
    current_user: User = Depends(require_permission("preferences:write")),
):
    svc = MemoryService(db, account_id=current_user.account_id)
    for key, value in req.model_dump().items():
        svc.set_preference(user_id=current_user.id, key=key, value=value, commit=False)
    db.commit()
    return serialize_message_response("偏好已保存")